requirements.txt

djangorestframework